# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FAA NOTAM: Developer Portal API",
    "description": "Backend REST API for the NOTAM Developer Portal",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhot:8080/api/ver{version}/",
      "description": "Local Development",
      "variables": {
        "version": {
          "default": "1",
          "description": "The API version number"
        }
      }
    }
  ],
  "paths": {
...
    "/notamApiToken/{id}": {
      "delete": {
        "tags": [
          "NotamApiToken"
        ],
        "summary": "Deletes API Token with the given Id.",
        "description": "<p>Allowed roles: Admin and Regular User can delete self API tokens only</p>",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The API Token Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The API Token Id.",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "NotamApiToken"
        ],
        "summary": "Retrieves API Token with the given Id.",
        "description": "<p>Allowed roles: Admin and Regular User</p>\r\n<p>Regular User can only access his/her own API Token</p>",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The API Token Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The API Token Id.",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotamApiTokenModel"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": [ ]
          }
        ]
      },
      "patch": {
        "tags": [
          "NotamApiToken"
        ],
        "summary": "Updates API Token with the given Id.",
        "description": "<p>Allowed roles: Admin and Regular User can update self API tokens only</p>",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The API Token Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The API Token Id.",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "The update API Token model.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotamApiTokenModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotamApiTokenModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotamApiTokenModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": [ ]
          }
        ]
      }
    },
    "/notamApiToken": {
      "get": {
        "tags": [
          "NotamApiToken"
        ],
        "summary": "Searches NOTAM API Tokens matching given criteria parameters.",
        "description": "<p>Allowed roles: Admin, Regular User</p>\r\n<p>For Regular User, the 'userId' parameter will be ignored and always set to current user's Id.</p>\r\n<p>\r\n            Provided parameters are just an example. Final parameter list will be defined during development phase based on the wireframes. (same for returned API Token fields)\r\n            </p>\r\n<p>All string fields will be matched using 'Contains' comparison. All non-null criterias should be AND-ed.</p>\r\n<p>Default and maximum 'limit' should be configurable in backend config.</p>",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "The owner user Id criteria.",
            "schema": {
              "type": "integer",
              "description": "The owner user Id criteria.",
              "format": "int32",
              "nullable": true
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "The user name criteria.",
            "schema": {
              "type": "string",
              "description": "The user name criteria.",
              "nullable": true
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "The name of the field to sort by.",
            "schema": {
              "type": "string",
              "description": "The name of the field to sort by.",
              "default": "Id",
              "nullable": true
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "description": "The sort order.",
            "schema": {
              "$ref": "#/components/schemas/SortOrder"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "The offset.",
            "schema": {
              "type": "integer",
              "description": "The offset.",
              "format": "int32",
              "default": 0,
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The limit.",
            "schema": {
              "type": "integer",
              "description": "The limit.",
              "format": "int32",
              "default": 20,
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotamApiTokenModelSearchResult"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "NotamApiToken"
        ],
        "summary": "Creates new NOTAM API token.",
        "description": "<p>Allowed roles: Admin, Regular User</p>\r\n<p>\r\n            The API Key should be auto-generated, e.g. generate new Guid and encode it to Base64 string.\r\n            </p>\r\n<p>\r\n            Set both IsEnabledByUser and IsEnabledByAdmin to True by default.\r\n            </p>",
        "requestBody": {
          "description": "The NOTAM API token details.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotamApiTokenModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotamApiTokenModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotamApiTokenModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotamApiTokenModel"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": [ ]
          }
        ]
      }
    },
    "/notamApiToken/{id}/enable": {
      "put": {
        "tags": [
          "NotamApiToken"
        ],
        "summary": "Enables NOTAM API Token with the given Id.",
        "description": "<p>Allowed roles: Admin, Regular User</p>\r\n<p>\r\n            If user is Admin, set 'IsEnabledByAdmin'=True. Admin can enable any API Token.\r\n            </p>\r\n<p>\r\n            If user is Regular User, set 'IsEnabledByUser'=True. Regular User can only enable his/her own API Token.\r\n            </p>",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The API Token Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The API Token Id.",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": [ ]
          }
        ]
      }
    },
    "/notamApiToken/{id}/disable": {
      "put": {
        "tags": [
          "NotamApiToken"
        ],
        "summary": "Disables NOTAM API Token with the given Id.",
        "description": "<p>Allowed roles: Admin, Regular User</p>\r\n<p>\r\n            If user is Admin, set 'IsEnabledByAdmin'=False. Admin can disable any API Token.\r\n            </p>\r\n<p>\r\n            If user is Regular User, set 'IsEnabledByUser'=False. Regular User can only disable his/her own API Token.\r\n            </p>",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The API Token Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The API Token Id.",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorModel": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Gets or sets the message.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "An API error Model."
      },
...
      "SortOrder": {
        "enum": [
          "Asc",
          "Desc"
        ],
        "type": "string",
        "description": "The sort order for search operations."
      },
...
      "NotamApiTokenModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The Id of the model.",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "The API token name.",
            "nullable": true,
            "example": "Development API token"
          },
          "key": {
            "type": "string",
            "description": "The API token key.",
            "nullable": true
          },
          "isEnabledByUser": {
            "type": "boolean",
            "description": "The flag indicating whether Token is enabled by User.",
            "example": true
          },
          "isEnabledByAdmin": {
            "type": "boolean",
            "description": "The flag indicating whether Token is enabled by Admin.",
            "example": false
          }
        },
        "additionalProperties": false,
        "description": "The NOTAM API Token model."
      },
      "UpdateNotamApiTokenModel": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The updated API token name.",
            "nullable": true,
            "example": "Updated development API token"
          }
        },
        "additionalProperties": false,
        "description": "The NOTAM API Token model."
      },
      "NotamApiTokenModelSearchResult": {
        "type": "object",
        "properties": {
          "totalCount": {
            "type": "integer",
            "description": "The total records count.",
            "format": "int32",
            "example": 124
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotamApiTokenModel"
            },
            "description": "The items.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "An model that represents search result."
      },
      "CreateNotamApiTokenModel": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The API token name.",
            "example": "Development API token"
          }
        },
        "additionalProperties": false,
        "description": "The Create NOTAM API Token model."
      },
...
    },
    "securitySchemes": {
      "jwt_bearer": {
        "type": "http",
        "description": "Authentication using JWT Token in header.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}